requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
click==7.1.2
Django==3.1.6
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2021.1
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
